requirements.txt
langchain>=0.2.0
langchain-openai>=0.2.0
langchain-google-genai>=0.2.0
langchain-community>=0.2.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
jupyter>=1.0.0
notebook>=7.0.0
ipykernel>=6.0.0
